move the config reconcile into the buffered merger factory
