+++ hudi-client/hudi-client-common/src/main/java/org/apache/hudi/table/HoodieTable.java
@@ -120,9 +120,7 @@ import java.util.stream.Stream;
 import static org.apache.hudi.avro.AvroSchemaUtils.resolveNullableSchema;
 import static org.apache.hudi.common.model.HoodieFailedWritesCleaningPolicy.EAGER;
 import static org.apache.hudi.common.model.HoodieFailedWritesCleaningPolicy.LAZY;
-import static org.apache.hudi.common.table.HoodieTableConfig.RECORD_MERGE_PROPERTY_PREFIX;
 import static org.apache.hudi.common.table.HoodieTableConfig.TABLE_METADATA_PARTITIONS;
-import static org.apache.hudi.common.util.ConfigUtils.extractWithPrefix;
 import static org.apache.hudi.common.util.StringUtils.EMPTY_STRING;
 import static org.apache.hudi.metadata.HoodieTableMetadataUtil.deleteMetadataPartition;
 import static org.apache.hudi.metadata.HoodieTableMetadataUtil.deleteMetadataTable;
@@ -164,9 +162,6 @@ public abstract class HoodieTable<T, I, K, O> implements Serializable {
     this.viewManager = getViewManager();
     this.metaClient = metaClient;
     this.taskContextSupplier = context.getTaskContextSupplier();
-
-    // Extract necessary properties from table config to write config.
-    reconcileWriteConfigs(metaClient, config);
   }
 
   protected HoodieTable(HoodieWriteConfig config, HoodieEngineContext context, HoodieTableMetaClient metaClient, FileSystemViewManager viewManager, TaskContextSupplier supplier) {
@@ -180,9 +175,6 @@ public abstract class HoodieTable<T, I, K, O> implements Serializable {
     this.viewManager = viewManager;
     this.metaClient = metaClient;
     this.taskContextSupplier = supplier;
-
-    // Extract necessary properties from table config to write config.
-    reconcileWriteConfigs(metaClient, config);
   }
 
   public boolean isMetadataTable() {
@@ -1294,14 +1286,5 @@ public abstract class HoodieTable<T, I, K, O> implements Serializable {
     // question: should we just return null when context is serialized as null? the mismatch reader context would throw anyway.
     return (ReaderContextFactory<T>) getContext().getReaderContextFactoryForWrite(metaClient, config.getRecordMerger().getRecordType(),
         config.getProps(), false);
-  }
-
-  private void reconcileWriteConfigs(HoodieTableMetaClient metaClient,
-                                     HoodieWriteConfig config) {
-    Map<String, String> mergeProperties = extractWithPrefix(
-        metaClient.getTableConfig().getProps(), RECORD_MERGE_PROPERTY_PREFIX);
-    if (!mergeProperties.isEmpty()) {
-      mergeProperties.forEach(config::setValue);
-    }
   }
 }
