+++ hudi-common/src/main/java/org/apache/hudi/common/util/ConfigUtils.java
@@ -52,6 +52,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static org.apache.hudi.common.config.HoodieReaderConfig.USE_NATIVE_HFILE_READER;
+import static org.apache.hudi.common.table.HoodieTableConfig.RECORD_MERGE_PROPERTY_PREFIX;
 import static org.apache.hudi.common.table.HoodieTableConfig.TABLE_CHECKSUM;
 import static org.apache.hudi.keygen.constant.KeyGeneratorOptions.KEYGENERATOR_CONSISTENT_LOGICAL_TIMESTAMP_ENABLED;
 
@@ -137,6 +138,19 @@ public class ConfigUtils {
     props.putIfAbsent(HoodiePayloadProps.PAYLOAD_ORDERING_FIELD_PROP_KEY, orderingFieldsAsString);
     props.putIfAbsent(HoodieTableConfig.ORDERING_FIELDS.key(), orderingFieldsAsString);
     return props;
+  }
+
+  /**
+   * Ensures that the prefixed merge properties are populated for mergers.
+   */
+  public static TypedProperties getMergeProps(TypedProperties props, TypedProperties tableProps) {
+    Map<String, String> mergeProps = extractWithPrefix(tableProps, RECORD_MERGE_PROPERTY_PREFIX);
+    if (mergeProps.isEmpty()) {
+      return props;
+    }
+    TypedProperties copied = TypedProperties.copy(props);
+    mergeProps.forEach(copied::setProperty);
+    return copied;
   }
 
   /**
